flask/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.3.2
flask-admin==1.5.2
flask-login-0.4.1
